requirements.txt
datasets
Pillow
tensorflow
ipykernel
matplotlib
ipywidgets